Refactor custom-button into a component and simplify
