+++ client/app/services/custom-button/custom-button.html
@@ -1,51 +1,26 @@
 <span>
-  <span ng-if="!vm.collapseCustomButtons">
-    <span ng-repeat="button in vm.customActions.buttons">
-       <button class="btn btn-default custom-button"
-             uib-tooltip="{{button.description}}"
-             tooltip-placement="bottom"
-             type="button"
-             ng-click="vm.customButtonAction(button, vm.serviceId, vm.serviceTemplateCatalogId)">
-         {{button.name }}
-       </button>
-    </span>
-    <div uib-dropdown class="btn-group" ng-repeat="buttonGroup in vm.customActions.button_groups">
-      <button uib-dropdown-toggle class="btn btn-default custom-button"
-        uib-tooltip="{{buttonGroup.description}}" tooltip-placement="bottom" type="button" id="{{buttonGroup.id}}">
-        {{ buttonGroup.name.split('|')[0] }}
-        <span class="caret"></span>
-      </button>
-      <ul uib-dropdown-menu role="menu" aria-labelledby="{{buttonGroup.id}}">
+  <div uib-dropdown class="btn-group custom-actions-menu">
+    <button uib-dropdown-toggle class="btn btn-default custom-button"
+      uib-tooltip="{{'Custom Actions'|translate}}" tooltip-placement="bottom" type="button" id="button_custom_actions">
+      {{'Custom Actions'|translate}}
+      <span class="caret"></span>
+    </button>
+    <ul uib-dropdown-menu role="menu" aria-labelledby="button_custom_actions">
+      <li ng-repeat="button in vm.customActions.buttons" role="presentation">
+        <a role="menuitem" tabindex="-1" class="custom-button-action"
+          ng-click="vm.invokeCustomAction(button)">
+          {{ button.name }}
+        </a>
+      </li>
+      <div ng-repeat="buttonGroup in vm.customActions.button_groups">
+        <li class="dropdown-header">{{ buttonGroup.name.split('|')[0] }}</li>
         <li ng-repeat="button in buttonGroup.buttons" role="presentation">
-          <a role="menuitem" tabindex="-1" ng-click="vm.customButtonAction(button, vm.serviceId, vm.serviceTemplateCatalogId)">
+          <a role="menuitem" tabindex="-1" class="custom-button-action"
+            ng-click="vm.invokeCustomAction(button)">
             {{ button.name }}
           </a>
         </li>
-      </ul>
-    </div>
-  </span>
-  <span ng-if="vm.collapseCustomButtons">
-    <div uib-dropdown class="btn-group custom-actions-menu">
-      <button uib-dropdown-toggle class="btn btn-default custom-button"
-        uib-tooltip="{{'Custom Actions'|translate}}" tooltip-placement="bottom" type="button" id="button_custom_actions">
-        {{'Custom Actions'|translate}}
-        <span class="caret"></span>
-      </button>
-      <ul uib-dropdown-menu role="menu" aria-labelledby="button_custom_actions">
-        <li ng-repeat="button in vm.customActions.buttons" role="presentation">
-          <a role="menuitem" class="custom-button-action" tabindex="-1" ng-click="vm.customButtonAction(button, vm.serviceId, vm.serviceTemplateCatalogId)">
-            {{ button.name }}
-          </a>
-        </li>
-        <div ng-repeat="buttonGroup in vm.customActions.button_groups">
-          <li class="dropdown-header">{{ buttonGroup.name.split('|')[0] }}</li>
-          <li ng-repeat="button in buttonGroup.buttons" role="presentation">
-            <a role="menuitem" tabindex="-1" class="custom-button-action" ng-click="vm.customButtonAction(button, vm.serviceId, vm.serviceTemplateCatalogId)">
-              {{ button.name }}
-            </a>
-          </li>
-        </div>
-      </ul>
-    </div>
-  </span>
+      </div>
+    </ul>
+  </div>
 </span>
