+++ client/app/services/service-details/service-details.html
@@ -13,10 +13,8 @@
     <div class="ss-details-header__actions">
       <div uib-dropdown class="ss-details-header__actions__inner dropdown-kebab-pf">
         <custom-button ng-if="vm.hasCustomButtons(vm.service)"
-                       service-template-catalog-id="vm.service.service_template.service_template_catalog_id"
                        service-id="vm.service.id"
-                       custom-actions="vm.service.custom_actions"
-                       actions="vm.service.actions">
+                       custom-actions="vm.service.custom_actions">
         </custom-button>
         <custom-dropdown class="custom-dropdown pull-left"
                          config="item"
